pyproject.toml
[project]
name = "40agent"
version = "0.1.0"
description = "AI Agent that can see and control VMs using Qwen3-omni multimodal model"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dashscope>=1.23.9",
    "pyaudio>=0.2.14",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=13.0",
    "pillow>=10.4.0",
    "opencv-python>=4.10.0.84",
    "numpy>=2.0.0",
    "libvirt-python>=10.9.0",
    "aiofiles>=24.1.0",
    "python-multipart>=0.0.12",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
